Tidy up registration and log out flows, finish user account feature

+++ src/lib/server/db/index.js
@@ -1,66 +1,68 @@
 import Database from 'better-sqlite3';
 import bcrypt from 'bcrypt'
+import { v4 as uuidv4 } from 'uuid';
 const saltRounds = 10;
 
 const db_path = './data/holidays.db';
 
 const db = new Database(db_path);
 
-export function getHolidays() {
-	const sql = `SELECT * FROM holidays`;
+export function getHolidays(userid) {
+	const sql = `SELECT * FROM holidays WHERE userid = $userid`;
 	const stmnt = db.prepare(sql);
-	const rows = stmnt.all();
+	const rows = stmnt.all({userid});
 	return rows;
 }
 
-export function addHoliday(from, to) {
-	const sql = `INSERT INTO holidays ('from', 'to') VALUES ($from, $to)`;
+export function addHoliday(userid, from, to) {
+	const sql = `INSERT INTO holidays ('userid', 'from', 'to') VALUES ($userid, $from, $to)`;
 	const stmnt = db.prepare(sql);
-	stmnt.run({ from, to });
+	stmnt.run({ userid, from, to });
 }
 
-export function deleteHoliday(id) {
-	const sql = `DELETE FROM holidays WHERE id = $id`;
+export function deleteHoliday(userid, id) {
+	const sql = `DELETE FROM holidays WHERE id = $id AND userid = $userid`;
 	const stmnt = db.prepare(sql);
-	stmnt.run({ id });
+	stmnt.run({ userid, id });
 }
 
-export function getAllowance() {
-	const sql = `SELECT allowance_days FROM config`;
+export function getAllowance(userid) {
+	const sql = `SELECT allowance_days FROM config WHERE userid = $userid`;
 	const stmnt = db.prepare(sql);
-	const res = stmnt.get();
+	const res = stmnt.get({ userid });
 	return res?.allowance_days ?? 0;
 }
 
-export function changeAllowance(newAllowance) {
-	const deleteSql = `DELETE FROM config`;
+export function changeAllowance(userid, newAllowance) {
+	const deleteSql = `DELETE FROM config WHERE userid= $userid`;
 	const deleteStmnt = db.prepare(deleteSql);
-	deleteStmnt.run();
+	deleteStmnt.run({ userid });
 
-	const insertSql = `INSERT INTO config (allowance_days) VALUES ($newAllowance)`;
+	const insertSql = `INSERT INTO config (userid, allowance_days) VALUES ($userid, $newAllowance)`;
 	const insertStmnt = db.prepare(insertSql);
-	insertStmnt.run({ newAllowance });
+	insertStmnt.run({ userid, newAllowance });
 }
 
-export function getExcludedDays() {
-	const sql = `SELECT * FROM excluded_days`;
+export function getExcludedDays(userid) {
+	const sql = `SELECT * FROM excluded_days WHERE userid = $userid`;
 	const stmnt = db.prepare(sql);
-	const rows = stmnt.all();
+	const rows = stmnt.all({ userid });
 	return rows;
 }
 
 export function createUser(username, password) {
 	const hash = bcrypt.hashSync(password, saltRounds)
-	const sql = `INSERT INTO users (username, password) VALUES ($username, $hash)`
+	const userid = uuidv4()
+	const sql = `INSERT INTO users (userid, username, password) VALUES ($userid, $username, $hash)`
 	const stmnt = db.prepare(sql)
-	stmnt.run({username, hash})
+	stmnt.run({ userid, username, hash })
 }
 
 export function authenticateUser(username, password) {
 	const sql = `SELECT * from users WHERE username=$username`
 	const stmnt = db.prepare(sql)
 	const res = stmnt.get({ username })
-	
+
 	if (!res) {
 		return false
 	}
@@ -82,6 +84,17 @@ export function getUserBySession(uuid) {
 	const sql = `SELECT * from users WHERE session = $uuid`
 	const stmnt = db.prepare(sql)
 	const res = stmnt.get({ uuid })
+	if (!res) {
+		return false
+	}
+
+	return res
+}
+
+export function getUserByUsername(username) {
+	const sql = `SELECT * from users WHERE username = $username`
+	const stmnt = db.prepare(sql)
+	const res = stmnt.get({ username })
 	if (!res) {
 		return false
 	}
